+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+# v0.2.0 (09-02-2023)
+- Added StopIfAbove method.
+- Added UpdateIfAbove method.
 # v0.1.0 (22-01-2023)
 - Added TimeLapse class.